chore(compute/metadata): fixup warnings
