+++ compute/metadata/metadata_test.go
@@ -16,10 +16,11 @@ package metadata
 
 import (
 	"bytes"
-	"io/ioutil"
+	"io"
 	"log"
 	"net/http"
 	"os"
+	"strings"
 	"sync"
 	"testing"
 )
@@ -106,7 +107,7 @@ type captureTransport struct {
 
 func (ct *captureTransport) RoundTrip(req *http.Request) (*http.Response, error) {
 	ct.url = req.URL.String()
-	return &http.Response{Body: ioutil.NopCloser(bytes.NewReader(nil))}, nil
+	return &http.Response{Body: io.NopCloser(io.Reader(bytes.NewReader(nil)))}, nil
 }
 
 type userAgentTransport struct {
@@ -125,5 +126,90 @@ type rrt struct {
 
 func (r *rrt) RoundTrip(req *http.Request) (*http.Response, error) {
 	r.gotUserAgent = req.Header.Get("User-Agent")
-	return &http.Response{Body: ioutil.NopCloser(bytes.NewReader(nil))}, nil
+	return &http.Response{Body: io.NopCloser(bytes.NewReader(nil))}, nil
+}
+
+func TestRetry(t *testing.T) {
+	tests := []struct {
+		name        string
+		timesToFail int
+		failCode    int
+		failErr     error
+		response    string
+		expectError bool
+	}{
+		{
+			name:     "no retries",
+			response: "test",
+		},
+		{
+			name:        "retry 500 once",
+			response:    "test",
+			failCode:    500,
+			timesToFail: 1,
+		},
+		{
+			name:        "retry io.ErrUnexpectedEOF once",
+			response:    "test",
+			failErr:     io.ErrUnexpectedEOF,
+			timesToFail: 1,
+		},
+		{
+			name:        "retry io.ErrUnexpectedEOF permanent",
+			failErr:     io.ErrUnexpectedEOF,
+			timesToFail: maxRetryAttempts + 1,
+			expectError: true,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			ft := &failingTransport{
+				timesToFail: tt.timesToFail,
+				failCode:    tt.failCode,
+				failErr:     tt.failErr,
+				response:    tt.response,
+			}
+			c := NewClient(&http.Client{Transport: ft})
+			s, err := c.Get("")
+			if tt.expectError && err == nil {
+				t.Fatalf("did not receive expected error")
+			} else if !tt.expectError && err != nil {
+				t.Fatalf("unexpected error: %v", err)
+			}
+
+			expectedCount := ft.failedAttempts + 1
+			if tt.expectError {
+				expectedCount = ft.failedAttempts
+			} else if s != tt.response {
+				// Responses are only meaningful if err == nil
+				t.Fatalf("c.Get() = %q, want %q", s, tt.response)
+			}
+
+			if ft.called != expectedCount {
+				t.Fatalf("failed %d times, want %d", ft.called, expectedCount)
+			}
+		})
+	}
+}
+
+type failingTransport struct {
+	timesToFail int
+	failCode    int
+	failErr     error
+	response    string
+
+	failedAttempts int
+	called         int
+}
+
+func (r *failingTransport) RoundTrip(req *http.Request) (*http.Response, error) {
+	r.called++
+	if r.failedAttempts < r.timesToFail {
+		r.failedAttempts++
+		if r.failErr != nil {
+			return nil, r.failErr
+		}
+		return &http.Response{StatusCode: r.failCode}, nil
+	}
+	return &http.Response{StatusCode: http.StatusOK, Body: io.NopCloser(strings.NewReader(r.response))}, nil
 }
